security has been improved

+++ private/controllers/Classes.php
@@ -1,51 +1,68 @@
 <?php
 
-/**
- * Classes controller
- */
 
+/**
+ * classes controller
+ */
 class Classes extends Controller
 {
 
     public function index()
     {
-
+        // code...
         if (!Auth::logged_in()) {
             $this->redirect('login');
         }
+
         $classes = new Classes_model();
 
         $school_id = Auth::getSchool_id();
-        $data = $classes->query("select * from classes where school_id = :school_id order by id desc", ['school_id' => $school_id]);
 
-        $crumbs[] = ['Dashboard', '/'];
+        if (Auth::access('admin')) {
+            $data = $classes->query("select * from classes where school_id = :school_id order by id desc", ['school_id' => $school_id]);
+        } else {
+
+            $class = new Classes_model();
+            $mytable = "class_students";
+            if (Auth::getRank() == "lecturer") {
+                $mytable = "class_lecturers";
+            }
+
+            $query = "select * from $mytable where user_id = :user_id && disabled = 0";
+            $arr['stud_classes'] = $class->query($query, ['user_id' => Auth::getUser_id()]);
+
+            $data = array();
+            if ($arr['stud_classes']) {
+                foreach ($arr['stud_classes'] as $key => $arow) {
+                    // code...
+                    $data[] = $class->first('class_id', $arow->class_id);
+                }
+            }
+        }
+
+        $crumbs[] = ['Dashboard', ''];
         $crumbs[] = ['Classes', 'classes'];
 
         $this->view('classes', [
             'crumbs' => $crumbs,
-            'rows' => $data,
-
+            'rows' => $data
         ]);
     }
 
     public function add()
     {
-
+        // code...
         if (!Auth::logged_in()) {
             $this->redirect('login');
         }
 
         $errors = array();
-
         if (count($_POST) > 0) {
 
-
             $classes = new Classes_model();
-
             if ($classes->validate($_POST)) {
 
-                $_POST['date'] = date('Y-m-d H:i:s');
-
+                $_POST['date'] = date("Y-m-d H:i:s");
 
                 $classes->insert($_POST);
                 $this->redirect('classes');
@@ -55,38 +72,30 @@ class Classes extends Controller
             }
         }
 
-
-        $crumbs[] = ['Dashboard', '/'];
+        $crumbs[] = ['Dashboard', ''];
         $crumbs[] = ['Classes', 'classes'];
         $crumbs[] = ['Add', 'classes/add'];
 
-
         $this->view('classes.add', [
             'errors' => $errors,
-            'crumbs' => $crumbs
+            'crumbs' => $crumbs,
 
         ]);
     }
 
     public function edit($id = null)
     {
-
+        // code...
         if (!Auth::logged_in()) {
             $this->redirect('login');
         }
 
-        $errors = array();
-
         $classes = new Classes_model();
 
-        if (count($_POST) > 0) {
-
-
-
+        $errors = array();
+        if (count($_POST) > 0 && Auth::access('lecturer') && Auth::i_own_content($row)) {
 
             if ($classes->validate($_POST)) {
-
-
 
                 $classes->update($id, $_POST);
                 $this->redirect('classes');
@@ -98,30 +107,35 @@ class Classes extends Controller
 
         $row = $classes->where('id', $id);
 
-        $crumbs[] = ['Dashboard', '/'];
+        $crumbs[] = ['Dashboard', ''];
         $crumbs[] = ['Classes', 'classes'];
         $crumbs[] = ['Edit', 'classes/edit'];
 
-        $this->view('classes.edit', [
-            'row' => $row,
-            'errors' => $errors,
-            'crumbs' => $crumbs
+        if (Auth::access('lecturer') && Auth::i_own_content($row)) {
 
-        ]);
+            $this->view('classes.edit', [
+                'row' => $row,
+                'errors' => $errors,
+                'crumbs' => $crumbs,
+            ]);
+        } else {
+            $this->view('access-denied');
+        }
     }
 
     public function delete($id = null)
     {
-
+        // code...
         if (!Auth::logged_in()) {
             $this->redirect('login');
         }
 
-        $errors = array();
 
         $classes = new Classes_model();
 
-        if (count($_POST) > 0) {
+        $errors = array();
+
+        if (count($_POST) > 0 && Auth::access('lecturer') && Auth::i_own_content($row)) {
 
             $classes->delete($id);
             $this->redirect('classes');
@@ -129,14 +143,18 @@ class Classes extends Controller
 
         $row = $classes->where('id', $id);
 
-        $crumbs[] = ['Dashboard', '/'];
+        $crumbs[] = ['Dashboard', ''];
         $crumbs[] = ['Classes', 'classes'];
         $crumbs[] = ['Delete', 'classes/delete'];
 
-        $this->view('classes.delete', [
-            'row' => $row,
-            'crumbs' => $crumbs
+        if (Auth::access('lecturer') && Auth::i_own_content($row)) {
 
-        ]);
+            $this->view('classes.delete', [
+                'row' => $row,
+                'crumbs' => $crumbs,
+            ]);
+        } else {
+            $this->view('access-denied');
+        }
     }
 }
